ref: Move mute utils in their own helper.
